Agregado de clases de DB

+++ Dominio/Categoria.cs
@@ -8,7 +8,7 @@ namespace Dominio
 {
     public class Categoria
     {
-        public int IDCategoria { get; set; }
+        public long IDCategoria { get; set; }
         public string Nombre { get; set; }
         public bool Estado { get; set; }
         public override string ToString()
