+++ Dominio/Usuario.cs
@@ -8,7 +8,7 @@ namespace Dominio
 {
     public class Usuario
     {
-        public int IDUsuario { get; set; }
+        public long IDUsuario { get; set; }
         public TipoUsuario TipoUser { get; set; }
         public string DNI { get; set; }
         public string Nombre { get; set; }
